Add test for CWG555
